requirement.txt
pymupdf==1.24.10
langchain==0.3.0
langchain-community==0.3.0
langchain-huggingface==0.3.0
langchain-chroma==0.3.0
langchain-openai==0.3.0
diskcache==5.6.3
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
gradio==4.44.0